book recommand

+++ SunReadSource/src/main/java/com/syzton/sunread/service/book/BookRepositoryService.java
@@ -2,12 +2,13 @@ package com.syzton.sunread.service.book;
 
 import com.syzton.sunread.assembler.book.BookAssembler;
 import com.syzton.sunread.dto.book.BookDTO;
+import com.syzton.sunread.exception.common.DuplicateException;
+import com.syzton.sunread.exception.common.NotFoundException;
 import com.syzton.sunread.model.book.Book;
-
 import com.syzton.sunread.model.book.Category;
 import com.syzton.sunread.repository.book.BookRepository;
 import com.syzton.sunread.repository.book.CategoryRepository;
-import javassist.NotFoundException;
+import com.syzton.sunread.repository.book.predicates.BookPredicates;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,12 +41,18 @@ public class BookRepositoryService implements BookService {
     }
 
 
-
+    @Transactional(rollbackFor = {DuplicateException.class})
     @Override
     public BookDTO add(BookDTO bookDTO) {
+
+        Book exits = bookRepo.findByIsbn(bookDTO.getIsbn());
+        if(exits != null){
+            throw new DuplicateException("book with isbn: "+bookDTO.getIsbn()+" is already exits..");
+        }
+
         BookAssembler assembler = new BookAssembler();
 
-        Book book =  assembler.fromDTO(bookDTO,categoryRepo);
+        Book book =  assembler.fromDTOtoEntireBook(bookDTO, categoryRepo);
 
         LOGGER.debug(book.toString());
 
@@ -54,9 +61,8 @@ public class BookRepositoryService implements BookService {
         return bookDTO;
     }
 
-    @Transactional(readOnly = true, rollbackFor = {NotFoundException.class})
     @Override
-    public Book findById(Long id) throws NotFoundException {
+    public Book findById(Long id) {
         LOGGER.debug("Finding a book entry with id: {}", id);
 
         Book found = bookRepo.findOne(id);
@@ -71,19 +77,20 @@ public class BookRepositoryService implements BookService {
 
     @Transactional(rollbackFor = {NotFoundException.class})
     @Override
-    public Book deleteById(Long id) throws NotFoundException {
+    public Book deleteById(Long id){
         LOGGER.debug("Deleting a to-do entry with id: {}", id);
 
         Book deleted = findById(id);
         LOGGER.debug("Deleting to-do entry: {}", deleted);
+        if(deleted == null)
+            throw new NotFoundException("No book found with id: " + id);
 
         bookRepo.delete(deleted);
         return deleted;
     }
     
-    @Transactional(rollbackFor = {NotFoundException.class})
     @Override
-    public Page<Book> findAll(Pageable pageable) throws NotFoundException{
+    public Page<Book> findAll(Pageable pageable){
 
         Page<Book> bookPages = bookRepo.findAll(pageable);
 
@@ -93,21 +100,31 @@ public class BookRepositoryService implements BookService {
 
     @Transactional(rollbackFor = {NotFoundException.class})
     @Override
-    public Page<Book> findByCategories(Set<Long> categoriyIds,Pageable pageable) throws NotFoundException{
+    public Page<Book> findByCategories(Set<Long> categoryIds,Pageable pageable){
 
         Set<Category> categories = new HashSet<>();
 
-        for(Long id : categoriyIds)
+        for(Long id : categoryIds)
         {
             Category category = categoryRepo.findOne(id);
             if(category == null){
                 throw new NotFoundException("no category be found wiht id: "+id+"");
             }
-            categories.add(categoryRepo.findOne(id));
+            categories.add(category);
         }
         Page<Book> bookPages = bookRepo.findByCategories(categories,pageable);
 
         return bookPages;
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public Page<Book> quickSearch(String searchTerm, Pageable pageable) {
+
+        Page<Book> bookPage = bookRepo.findAll(BookPredicates.quickSearchContains(searchTerm), pageable);
+        if(bookPage==null)
+            throw new NotFoundException("no satisfy book result with this search term : "+searchTerm);
+        return bookPage;
     }
